+++ manager.js
@@ -10,6 +10,18 @@ const { manager }  = require('@absolunet/manager');
 const paths        = require('./dist/node/helpers/paths');
 
 
+const getPipelineStep = (name, version) => {
+	return {
+		step: {
+			name:   `Test ${name} Node.js version`,
+			image:  `node:${version}`,
+			caches: ['node'],
+			script: ['npm install', 'npm test']
+		}
+	};
+};
+
+
 manager.init({
 	repositoryType: 'single-package',
 	dist: {
@@ -21,25 +33,42 @@ manager.init({
 				terminal.print(`Update Node version in package.json / .travis.yml / bitbucket-pipelines.yml`).spacer();
 
 				const latest = await nodejsLatest.latest();
-				const { version } = semver.coerce(semver.major(latest.version));
+				const { version: latestVersion } = semver.coerce(semver.major(latest.version));
+				const stableVersion = semver.major(latestVersion) % 2 === 0 ? latestVersion : semver.coerce(semver.major(latestVersion) - 1).version;
 
+
+				//-- package.json
 				const packageFile = `${paths.root}/package.json`;
 				const packageData = fss.readJson(packageFile);
-				packageData.engines.node = `>= ${version}`;
+				packageData.engines.node = `>= ${stableVersion}`;
 				fss.writeJson(packageFile, packageData, { space: 2 });
 
+
+				//-- .travis.yml
 				const travisFile = `${paths.root}/.travis.yml`;
 				const travisData = fss.readYaml(travisFile);
-				travisData.node_js[1] = version;
+				travisData.node_js = ['node', latestVersion];  // eslint-disable-line camelcase
+
+				if (latestVersion !== stableVersion) {
+					travisData.node_js.push(stableVersion);
+				}
+
 				fss.writeYaml(travisFile, travisData);
 
+
+				//-- bitbucket-pipelines.yml
 				const pipelinesFile = `${paths.root}/bitbucket-pipelines.yml`;
 				const pipelinesData = fss.readYaml(pipelinesFile);
-				pipelinesData.pipelines.default[0].parallel[1].step.image = `node:${version}`;
+				pipelinesData.pipelines.default[0].parallel = [getPipelineStep('latest', 'latest'), getPipelineStep('original latest', latestVersion)];
+
+				if (latestVersion !== stableVersion) {
+					pipelinesData.pipelines.default[0].parallel.push(getPipelineStep('original latest stable', stableVersion));
+				}
+
 				fss.writeYaml(pipelinesFile, pipelinesData);
 
 
-
+				//-- license
 				terminal.print(`Update year in license`).spacer();
 				const licenseFile = `${paths.root}/license`;
 				let licenseData = fss.readFile(licenseFile, 'utf8');